Clean up JS (#7)
