+++ ide/package.json
@@ -90,8 +90,8 @@
     "build": "node build.cjs",
     "watch": "node build.cjs -w",
     "tc": "tsc",
-    "pretest": "npm run tc && npm run build",
-    "test": "node ./out/tests/runTests.js"
+    "pretest": "tsc --outDir test && npm run build",
+    "test": "node ./test/tests/runTests.js"
   },
   "devDependencies": {
     "@types/adm-zip": "^0.4.34",
